+++ src/components/Styles/selectpats.css
@@ -1,0 +1,10 @@
+.selectpats {
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100vh;
+    background-color: rgba(255, 255, 255);
+    display: flex;
+    justify-content: center;
+    align-items: center;
+  }
